Root redirect: check the logged in/out state when handling the / route, not when installing it
Should fix `wp-desktop` test failures that are related to login and logout.

If the user was logged in at the time when the `/` route handler was installed, then the `/`
route won't be handled well right after the user logs out. And vice versa during the logged-out
to logged-in transition.

The fix moves the logged in/out check inside the route handler, to be executed every time.

The Desktop app doesn't reboot/reload during these transitions, unlike the Web app, that's why
the transitions fail.